EZP-24765: Give possibility to set additional paths for a given theme
> https://jira.ez.no/browse/EZP-27465

This patch makes it possible to add arbitrary Twig template directories to a theme from configuration.
This can be useful when you want to define templates from third-party bundles as part of one of your themes,
or when upgrading your application in order to use eZ Platform design engine, and that your existing templates
are not yet following the convention.

Practical use case would be to define a `standard` theme for templates provided by the kernel.

```yaml
ezdesign:
    design_list:
        my_design: [my_theme, some_other_theme]
    templates_theme_paths:
        # FOSUserBundle templates will be part of "my_theme" theme
        my_theme:
            - '%kernel.root_dir%/../vendor/friendsofsymfony/user-bundle/Resources/views'
```