Fix Twig loading exception on first page load
> https://jira.ez.no/browse/EZP-27510
> https://github.com/ezsystems/ezpublish-kernel/pull/2028

This patch fixes the Twig exception saying it can't load a given template
using `@standard` namespace.

This is due to [EZP-27510](https://jira.ez.no/browse/EZP-27510) as ConfigResolver's
default scope is not up-to-date after cache warmup process.
This fix requires https://github.com/ezsystems/ezpublish-kernel/pull/2028 to be merged.

Note that I replaced `$design$` parameter injection by the whole ConfigResolver.
This is to ensure that we'll have the correct design when we need it.